requirements.txt
pyyaml
jsonschema
markdown-it-py
pytest
pre-commit
requests
